iot2050: Split the build for pg1 and pg2
Since the pg1 and the pg2 boards can not use the same SPL anymore, it's
not possbile to build a single flash.bin for both variants.

A new config is added to indicates the build is for pg2 board. Hence
now the pg1 and the pg2 boards have separated defconfig files.

The runtime board_is_sr1() check does make no sence anymore, so remove
it and replace with build time checking.

Signed-off-by: Su Baocheng <baocheng.su@siemens.com>